V1.04.03
changes on connot resolve problem and add search bar for find the friend

+++ app/src/main/java/com/pm/appdev/duta/chats/MessagesAdapter.java
@@ -3,7 +3,6 @@ package com.pm.appdev.duta.chats;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
-//import android.view.ActionMode;
 import androidx.appcompat.view.ActionMode;
 import android.view.LayoutInflater;
 import android.view.Menu;
@@ -14,7 +13,6 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -23,8 +21,6 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.pm.appdev.duta.R;
 import com.pm.appdev.duta.Common.Constants;
-import com.pm.appdev.duta.selectfriend.SelectFriendActivity;
-//import com.pm.appdev.duta.;
 import com.bumptech.glide.Glide;
 import com.google.firebase.auth.FirebaseAuth;
 
@@ -39,7 +35,7 @@ public class MessagesAdapter extends RecyclerView.Adapter<MessagesAdapter.Messag
     private FirebaseAuth firebaseAuth;
 
     private ActionMode actionMode;
-    private  ConstraintLayout selectedView;
+    private ConstraintLayout selectedView;
 
     public MessagesAdapter(Context context, List<MessageModel> messageList) {
         this.context = context;
@@ -58,25 +54,22 @@ public class MessagesAdapter extends RecyclerView.Adapter<MessagesAdapter.Messag
 
         MessageModel message = messageList.get(position);
         firebaseAuth = FirebaseAuth.getInstance();
-        String currentUserId= firebaseAuth.getCurrentUser().getUid();
+        String currentUserId = firebaseAuth.getCurrentUser().getUid();
 
         String fromUserId = message.getMessageFrom();
 
         SimpleDateFormat sfd = new SimpleDateFormat("dd-MM-yyyy HH:mm");
 
         String dateTime = sfd.format(new Date(message.getMessageTime()));
-        String [] splitString = dateTime.split(" ");
+        String[] splitString = dateTime.split(" ");
         String messageTime = splitString[1];
 
-        if(fromUserId.equals(currentUserId)){
+        if (fromUserId.equals(currentUserId)) {
 
-            if(message.getMessageType().equals(Constants.MESSAGE_TYPE_TEXT))
-            {
+            if (message.getMessageType().equals(Constants.MESSAGE_TYPE_TEXT)) {
                 holder.llSent.setVisibility(View.VISIBLE);
                 holder.llSentImage.setVisibility(View.GONE);
-            }
-            else
-            {
+            } else {
                 holder.llSent.setVisibility(View.GONE);
                 holder.llSentImage.setVisibility(View.VISIBLE);
             }
@@ -91,15 +84,11 @@ public class MessagesAdapter extends RecyclerView.Adapter<MessagesAdapter.Messag
                     .load(message.getMessage())
                     .placeholder(R.drawable.ic_image)
                     .into(holder.ivSent);
-        }
-        else
-        {
-            if(message.getMessageType().equals(Constants.MESSAGE_TYPE_TEXT)) {
+        } else {
+            if (message.getMessageType().equals(Constants.MESSAGE_TYPE_TEXT)) {
                 holder.llReceived.setVisibility(View.VISIBLE);
                 holder.llReceivedImage.setVisibility(View.GONE);
-            }
-            else
-            {
+            } else {
                 holder.llReceived.setVisibility(View.GONE);
                 holder.llReceivedImage.setVisibility(View.VISIBLE);
             }
@@ -126,14 +115,12 @@ public class MessagesAdapter extends RecyclerView.Adapter<MessagesAdapter.Messag
             public void onClick(View view) {
                 String messageType = view.getTag(R.id.TAG_MESSAGE_TYPE).toString();
                 Uri uri = Uri.parse(view.getTag(R.id.TAG_MESSAGE).toString());
-                if(messageType.equals(Constants.MESSAGE_TYPE_VIDEO))
-                {
-                    Intent intent = new Intent(Intent.ACTION_VIEW,uri);
+                if (messageType.equals(Constants.MESSAGE_TYPE_VIDEO)) {
+                    Intent intent = new Intent(Intent.ACTION_VIEW, uri);
                     intent.setDataAndType(uri, "video/mp4");
                     context.startActivity(intent);
-                }
-                else if(messageType.equals(Constants.MESSAGE_TYPE_IMAGE)){
-                    Intent intent = new Intent(Intent.ACTION_VIEW,uri);
+                } else if (messageType.equals(Constants.MESSAGE_TYPE_IMAGE)) {
+                    Intent intent = new Intent(Intent.ACTION_VIEW, uri);
                     intent.setDataAndType(uri, "image/jpg");
                     context.startActivity(intent);
                 }
@@ -144,16 +131,16 @@ public class MessagesAdapter extends RecyclerView.Adapter<MessagesAdapter.Messag
         holder.clMessage.setOnLongClickListener(new View.OnLongClickListener() {
             @Override
             public boolean onLongClick(View view) {
-                if(actionMode!=null)
+                if (actionMode != null)
                     return false;
 
                 selectedView = holder.clMessage;
 
-                actionMode = ((AppCompatActivity)context).startSupportActionMode(actionModeCallBack);
+                actionMode = ((AppCompatActivity) context).startSupportActionMode(actionModeCallBack);
 
                 holder.clMessage.setBackgroundColor(context.getResources().getColor(R.color.colorAccent));
 
-                return  true;
+                return true;
             }
         });
     }
@@ -163,7 +150,7 @@ public class MessagesAdapter extends RecyclerView.Adapter<MessagesAdapter.Messag
         return messageList.size();
     }
 
-    public class MessageViewHolder  extends  RecyclerView.ViewHolder{
+    public class MessageViewHolder extends RecyclerView.ViewHolder {
 
         private LinearLayout llSent, llReceived, llSentImage, llReceivedImage;
         private TextView tvSentMessage, tvSentMessageTime, tvReceivedMessage, tvReceivedMessageTime;
@@ -186,8 +173,8 @@ public class MessagesAdapter extends RecyclerView.Adapter<MessagesAdapter.Messag
 
             llSentImage = itemView.findViewById(R.id.llSentImage);
             llReceivedImage = itemView.findViewById(R.id.llReceivedImage);
-            ivSent =itemView.findViewById(R.id.ivSent);
-            ivReceived =itemView.findViewById(R.id.ivReceived);
+            ivSent = itemView.findViewById(R.id.ivSent);
+            ivReceived = itemView.findViewById(R.id.ivReceived);
 
             tvImageSentTime = itemView.findViewById(R.id.tvSentImageTime);
             tvImageReceivedTime = itemView.findViewById(R.id.tvReceivedImageTime);
@@ -195,89 +182,72 @@ public class MessagesAdapter extends RecyclerView.Adapter<MessagesAdapter.Messag
         }
     }
 
-    public  ActionMode.Callback actionModeCallBack = new ActionMode.Callback() {
+    public ActionMode.Callback actionModeCallBack = new ActionMode.Callback() {
         @Override
         public boolean onCreateActionMode(ActionMode actionMode, Menu menu) {
-
+            // Load the menu resource like an object
             MenuInflater inflater = actionMode.getMenuInflater();
             inflater.inflate(R.menu.menu_chat_options, menu);
 
             String selectedMessageType = String.valueOf(selectedView.getTag(R.id.TAG_MESSAGE_TYPE));
-            if(selectedMessageType.equals(Constants.MESSAGE_TYPE_TEXT))
-            {
+            if (selectedMessageType.equals(Constants.MESSAGE_TYPE_TEXT)) {
                 MenuItem itemDownload = menu.findItem(R.id.mnuDownload);
                 itemDownload.setVisible(false);
             }
             return true;
         }
 
-        @Override;
+        @Override
         public boolean onPrepareActionMode(ActionMode actionMode, Menu menu) {
             return false;
         }
 
         @Override
         public boolean onActionItemClicked(ActionMode actionMode, MenuItem menuItem) {
-
             String selectedMessageId = String.valueOf(selectedView.getTag(R.id.TAG_MESSAGE_ID));
             String selectedMessage = String.valueOf(selectedView.getTag(R.id.TAG_MESSAGE));
             String selectedMessageType = String.valueOf(selectedView.getTag(R.id.TAG_MESSAGE_TYPE));
 
             int itemId = menuItem.getItemId();
-            switch (itemId)
-            {
-                case  R.id.mnuDelete:
 
-                    if(context instanceof  ChatActivity)
-                    {
-                        ((ChatActivity)context).deleteMessage(selectedMessageId, selectedMessageType);
+            // Replace switch with if-else to avoid "Constant expression required" error
+            if (itemId == R.id.mnuDelete) {
+                if (context instanceof ChatActivity) {
+                    ((ChatActivity) context).deleteMessage(selectedMessageId, selectedMessageType);
+                }
+                actionMode.finish();
+            } else if (itemId == R.id.mnuDownload) {
+                if (context instanceof ChatActivity) {
+                    ((ChatActivity) context).downloadFile(selectedMessageId, selectedMessageType, false);
+                }
+                actionMode.finish();
+            } else if (itemId == R.id.mnuShare) {
+                if (selectedMessageType.equals(Constants.MESSAGE_TYPE_TEXT)) {
+                    Intent intentShare = new Intent();
+                    intentShare.setAction(Intent.ACTION_SEND);
+                    intentShare.putExtra(Intent.EXTRA_TEXT, selectedMessage);
+                    intentShare.setType("text/plain");
+                    context.startActivity(intentShare);
+                } else {
+                    if (context instanceof ChatActivity) {
+                        ((ChatActivity) context).downloadFile(selectedMessageId, selectedMessageType, true);
                     }
-                    actionMode.finish();
-                    break;
-                case  R.id.mnuDownload:
-                    if(context instanceof  ChatActivity)
-                    {
-                        ((ChatActivity)context).downloadFile(selectedMessageId, selectedMessageType, false);
-                    }
-                    actionMode.finish();
-                    break;
-                case  R.id.mnuShare:
-                    if(selectedMessageType.equals(Constants.MESSAGE_TYPE_TEXT)){
-                        Intent intentShare = new Intent();
-                        intentShare.setAction(Intent.ACTION_SEND);
-                        intentShare.putExtra(Intent.EXTRA_TEXT, selectedMessage);
-                        intentShare.setType("text/plain");
-                        context.startActivity(intentShare);
-                    }
-                    else
-                    {
-                        if(context instanceof  ChatActivity)
-                        {
-                            ((ChatActivity)context).downloadFile(selectedMessageId, selectedMessageType, true);
-                        }
-                    }
-                    actionMode.finish();
-                    break;
-                case  R.id.mnuForward:
-                    R.menu.menu_chat_options;
-
-                    if(context instanceof  ChatActivity)
-                    {
-                        ((ChatActivity) context).forwardMessage(selectedMessageId, selectedMessage, selectedMessageType);
-                    }
-
-                    actionMode.finish();
-                    break;
+                }
+                actionMode.finish();
+            } else if (itemId == R.id.mnuForward) {
+                if (context instanceof ChatActivity) {
+                    ((ChatActivity) context).forwardMessage(selectedMessageId, selectedMessage, selectedMessageType);
+                }
+                actionMode.finish();
             }
+
             return false;
         }
 
         @Override
         public void onDestroyActionMode(ActionMode actionMode) {
-            actionMode =null;
+            actionMode = null;
             selectedView.setBackgroundColor(context.getResources().getColor(R.color.chat_background));
-
         }
     };
-
 }
